Added a base agent class
